refactor: migrate hooks (use-additional-editor-mention) from web/app/ce to web/app/core
